requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-jose[cryptography]
passlib[bcrypt]
pytesseract
opencv-python-headless
groq
google-generativeai
matplotlib
graphviz
python-dotenv
google-auth
psycopg2-binary
email-validator
python-multipart

